Estuary OSD showing wrong value for ‘Next’ programme when playing radio.
Change the order of the tests used to establish what type of media is playing when determining how to set OSDNextLabelVar. 

Variables.xml currently sets the value of OSDNextLabelVar by checking if the ‘Visualisation” window is active before checking whether we’re playing ‘livetv’. 

Any radio shows played through the PVR pass the ‘livetv’ test, and can correctly show the Start Time, End Time and Name of the next programme.  This currently fails however, because the ‘Visualisation’ check is successfully satisfied first.